% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova2}
\alias{anova2}
\title{Anova for Linear Regression}
\usage{
anova2(formula, data, subset, na.action = "omit", intercept = TRUE)
}
\arguments{
\item{formula}{A framework for how should the linear regression model be given.}

\item{data}{A matrix of (nxp) whose data will be used to form the model.}

\item{subset}{The conditions of which rows in data will be used.}

\item{na.action}{How will the missing values be handled. It is defauled to "omit" value.}

\item{intercept}{A boolean value if the linear regression model will have an
intercept or not. It is defaulted to TRUE.}
}
\value{
A matrix that contains degrees of freedom, sum of squares, mean sum of
squares, f statistics, p values for each of the covariates in the fitted
linear regression model.
}
\description{
Performs ANOVA of the linear regression model
}
\examples{
data = read.csv("data/melb_data.csv")
output = anova2(data$Price~data$Bedroom2+data$Bathroom+data$Landsize,data,
na.action="omit",intercept=TRUE)
ss = output[,2]
f = output[,4]
}
