% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model_performance}
\alias{evaluate_model_performance}
\title{Evaluate Model Performance}
\usage{
evaluate_model_performance(fit_result, dataset, response_var, predictor_vars)
}
\arguments{
\item{fit_result}{A list containing the results from a linear regression model}

\item{dataset}{The dataset used to fit the model}

\item{response_var}{The name of the response variable}

\item{predictor_vars}{A vector of predictor variables}
}
\value{
A list containing:
\itemize{
\item R_squared: The R-squared value
\item Adj_R_squared: The Adjusted R-squared value
\item F_statistic: The F-statistic value
\item p_value_F: The p-value for the F-statistic
}
}
\description{
This function evaluates the performance of a linear regression model by computing R-squared,
Adjusted R-squared, F-statistic, and its associated p-value.
}
