% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_width_CI.R
\name{get_width_ci}
\alias{get_width_ci}
\title{Calculates width of uncertainty intervals}
\usage{
get_width_ci(est, iso_code, coverage)
}
\arguments{
\item{est}{tibble which contains mCPR estimates and uncertainty intervals. Columns: Country or area, iso, Year, Median, U95, L95, U80, L80}

\item{iso_code}{country iso code}

\item{coverage}{confidence intervals. Options are: 80 or 95}
}
\value{
tibble with year and corresponding width of uncertainty intervals
}
\description{
Calculates width of uncertainty intervals
}
\examples{
get_width_ci(est, iso_code = 4, coverage = 95)
}
