% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_regression_modify.R
\name{glm_regression}
\alias{glm_regression}
\title{glm regression}
\usage{
glm_regression(formula, data, family)
}
\arguments{
\item{formula}{math expression}

\item{data}{your data}

\item{family}{binomial/poisson}
}
\value{
formula,coefficients,ite,fitted.values,residual,var,sigma_2.hat,residual.std.error
}
\description{
glm regression
}
