\name{hw3-package}
\alias{hw3-package}
\alias{hw3}
\docType{package}
\title{
\packageTitle{hw3}
}
\description{
\packageDescription{hw3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hw3}
\packageIndices{hw3}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hw3}

Maintainer: \packageMaintainer{hw3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
