\name{JacobBML-package}
\alias{JacobBML-package}
\alias{JacobBML}
\docType{package}
\title{
Simple preliminary take on the BML traffic grid model}
\description{The BML traffic grid is a simple model that exhibits a phase change}
\details{
\tabular{ll}{
Package: \tab JacobBML\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-03\cr
License: \tab What license is it under?\cr
}
See the pdf I'm writing. Geeze}
\author{Jacob Teter

Maintainer: JET <jacob.teter@gmail.com>
}
\references{See the website: http://mae.ucdavis.edu/dsouza/bml.html for more than you'd ever want to know on BML traffic models}
\keyword{ package }
\seealso{}
\examples{}