% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear_model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{input response and variables}

\item{data}{input the data frame}
}
\value{
the estimated coefficients and statistics of t test and F test
}
\description{
Compute the estimated coefficients and statistics from linear models
}
\examples{
X <- c(151, 174, 138, 186, 128, 136, 179, 163, 152, 131)
Y <- c(63, 81, 56, 91, 47, 57, 76, 72, 62, 48)
data <- data.frame(Y,X)
linear_model(Y~X, data)

x <- cbind(mtcars$disp, mtcars$drat)
y <- mtcars$mpg
linear_model(y~x, mtcars)

}
