% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{linear regression}
\usage{
linear_regression(formula, data)
}
\arguments{
\item{formula}{the  expression of linear regression formula}

\item{data}{our data}
}
\value{
formula, residuals.descript,Coefficients,CI,residual.std.error,residual.df,R_squared,Adjusted.R_squared,F.value,df,pvalue.f,fittedvalues,residual.
}
\description{
linear regression
}
\examples{
example.r
}
