% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{logistic regression}
\usage{
logistic(formula, data)
}
\arguments{
\item{formula}{math expression}

\item{data}{your data}
}
\value{
formula,coefficients,ite,fitted.values,residual,var,sigma_2.hat,residual.std.error
}
\description{
logistic regression
}
