% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melb_data.R
\docType{data}
\name{melb_data}
\alias{melb_data}
\title{Melbourne Housing data}
\format{
13580 rows and 21 columns
}
\source{
\href{https://www.kaggle.com/code/alexisbcook/missing-values/data?select=melb_data.csv}
}
\usage{
melb_data
}
\description{
Data that was scraped from Domain.com.au weekly. This data give certain
details about the housing in Melbourne, such as "Rooms", "Price", etc.
This csv file contains 13580 rows and 21 columns.
}
\examples{
melb_data
}
\keyword{datasets}
