% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_analysis.R
\name{miceanalyze}
\alias{miceanalyze}
\title{analysis after imputation}
\usage{
miceanalyze(formula, imputed.data = imputed.data, n, n.imp = 5, mydata)
}
\arguments{
\item{formula}{the glm regression expression}

\item{imputed.data}{output of imputed data using MICE}

\item{n}{number of covariates}

\item{mydata}{original data}
}
\value{
t,vm,w.bar,B,m,r,lamda,RE
}
\description{
when we use mice to impute the data including NA, the next step we want to do is analyze the efficacy of imputation. RE is a crucial factor to assess the efficacy of imputation.
}
