% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_cor.R
\name{modified_cor}
\alias{modified_cor}
\title{modified_cor}
\usage{
modified_cor(x, y = NULL)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame with complete cases.}

\item{y}{NULL(default) or a numeric vector, matrix or data frame with complete cases and compatible dimensions to x.}
}
\value{
the pearson correlation between x,y or columns of x matrix
}
\description{
Gets the pearson correlation between two vectors or columns of matrices with complete cases.
}
\details{
A matrix/data frame for x or both x,y vectors/matrices should be given to calculate pearson correlation.
The inputs must be numeric or logical, without missing values. For modified_cor, pearson correlation will be calculated, which produces same results for function cor(x,y,use="all.obs", method="pearson")
The denominator n - 1 is used when computing unbiased estimators of the covariances for i.i.d. observations(when only one observation for x,y are given, the function will return NaN).
}
\examples{
## For complete matrices and vectors with compatible dimensions.
## x,y are both vectors with one element
modified_cor(1,2) # NaN

## x,y are both vectors
modified_cor(c(1,2,3,TRUE,FALSE),c(FALSE,TRUE,4,7,0)) # 0.3019786
## y is NULL, x is matrix
x<-matrix(runif(1000,min=0,max=100),10,100)
stopifnot(all.equal(cor(x,y=NULL,use="all.obs",method="pearson"), modified_cor(x)))

## Both x,y are given
x1<-matrix(runif(100,min=0,max=100),10,10)
y1<-c(runif(10,min=0,max=100))
stopifnot(all.equal(cor(x1,y1,use="all.obs",method="pearson"), modified_cor(x1,y1)))


## Consider error conditions:
## For complete matrices and vectors with incompatible dimensions
modified_cor(matrix(c(1,2,3,4,3,4,5,6),2,4),c(1,2,3,4)) # Error: incompatible dimensions

## For objects with missing values
x3<-matrix(runif(5000,min=0,max=100),20,50)
x3[15,25]=NA
modified_cor(x3) # Error: missing observations in cov/cor

## For objects not numeric/logical/atomic
modified_cor(c(1,2,TRUE,"abc"),c(1,2,3,4)) # Error: ((is.numeric(x) || is.logical(x)) & is.atomic(x)) is not TRUE

}
\seealso{
modified_cov for covariance computation for complete vectors/matrices.
}
