% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_cov.R
\name{modified_cov}
\alias{modified_cov}
\title{modified_cov}
\usage{
modified_cov(x, y = NULL, use = "all.obs", method = "pearson")
}
\arguments{
\item{x, y}{input value}
}
\value{
the covariances between x,y or columns of x matrix
}
\description{
Gets the covariance between two vectors or columns of matrices with complete cases under the method option "pearson".
}
\examples{
modified_cov(c(1,2,3,4),c(2,3,4,5))

}
