% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova2}
\alias{anova2}
\title{Anova for Linear Regression}
\usage{
anova2(formula,data,subset,na.action)
}
\arguments{
\item{formula}{A framework for how should the linear regression model be given.}

\item{data}{A matrix of (nxp) whose data will be used to form the model.}

\item{na.action}{How will the missing values be handled. It is defauled to "omit" value.}
}
\value{
A matrix that contains degrees of freedom, sum of squares, mean sum of
squares, f statistics, p values for each of the covariates in the fitted
linear regression model.
}
\description{
Performs ANOVA of the linear regression model
}
\examples{
data = read.csv("data/melb_data.csv")
output = anova2(data$Price~data$Bedroom2+data$Bathroom+data$Landsize,data,
na.action="omit")
ss = output[,2]
f = output[,4]

}
