% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm2}
\alias{lm2}
\title{Linear Regression and Hypothesis Testing}
\usage{
lm2(formula,data,na.action)
}
\arguments{
\item{formula}{A framework for how should the linear regression model be given.}

\item{data}{A matrix of (nxp) whose data will be used to form the model.}

\item{na.action}{How will the missing values be handled. It is defaulted to "omit" value.}
}
\value{
A list that contains the following for the fitted linear regression model:
coefficients of the linear regression model, residuals, sigma, coefficient
of determination, adjusted coefficient of determination, unscaled covariance,
f statistic, variance-covariance matrix, SSE, SSR, SSY, design matrix, and
the response
}
\description{
Preprocess the data, fits a specified linear regression model, and performs
basic hypothesis testing. It is designed to copy some of the values from the
summary of the lm().
}
\examples{
output = lm2(melb_data$Price~melb_data$Bedroom2+melb_data$Bathroom+melb_data$Landsize,melb_data,
na.action="omit")
betas = output$coefficients
res = output$residuals

}
