% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_cor.R
\name{modified_cor}
\alias{modified_cor}
\title{modified_cor}
\usage{
modified_cor(x, y = NULL)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame with complete cases.}

\item{y}{NULL(default) or a numeric vector, matrix or data frame with complete cases and compatible dimensions to \code{x}.}
}
\value{
the pearson correlation between x,y or columns of x matrix
}
\description{
Gets the pearson correlation between two vectors or columns of matrices with complete cases.
}
\details{
A matrix/data frame for \code{x} or both \code{x},\code{y} vectors/matrices should be given to calculate pearson correlation.
The inputs must be numeric or logical, without missing values. For \code{modified_cor()}, pearson correlation will be calculated, which produces same results for function \code{cor(x,y,use="all.obs", method="pearson")}
The denominator \code{n-1} is used when computing unbiased estimators of the covariances for i.i.d. observations(when only one observation for both \code{x}, \code{y} are given, the function will return \code{NaN}).
}
\note{
Pearson method performs better on normal distributed samples.While Spearman and Kendall methods are usually applied on samples that are not in normal distribution.
}
\examples{
## For complete matrices and vectors with compatible dimensions.
## x,y are both vectors with one element
modified_cor(1,2) # NaN

## x,y are both vectors
modified_cor(c(1,2,3,TRUE,FALSE),c(FALSE,TRUE,4,7,0)) # 0.3019786

## One of x,y is a matrix, and the other is a vector
x1<-matrix(rnorm(500,0,1),50,10)
y1<-c(rnorm(50,0,1))
stopifnot(all.equal(cor(x1,y1,use="all.obs",method="pearson"), modified_cor(x1,y1)))


}
\seealso{
\code{\link[=modified_cov]{modified_cov()}} for covariance computation for complete vectors/matrices.
}
